% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFADN.R
\name{readFADN}
\alias{readFADN}
\title{readFADN}
\usage{
readFADN(start = 1990, end = 2009, int_rate = 1.01,
  data_ver = "20150722")
}
\arguments{
\item{start}{First year for which data needs to be read in from FADN files}

\item{end}{last year for which data needs to be read in from FADN files}

\item{int_rate}{interest rate+100 percent. Here, 1.01 stands for 1 percent interest rate. Can be changed for sensitivity analysis.}

\item{data_ver}{the data version for FADN data.}
}
\value{
data frame.
}
\description{
reads a harmonized FADN and EOBS(ECA&D) data. (Link To be provided to download locations)
}
\examples{
\dontrun{
x <- readFADN(start=1990, end=2009, int_rate=1.01,data_ver="20150722")
}
}
\author{
Abhijeet Mishra
}
